#Many fixed and rewrites later. Money is now a ValueObject, Strategy pattern is better implemented, Statistics page works, everything from top to bottom cleaned up as to not be broken

+++ FysioEnterprise.Domain/ValueObjects/Price.cs
@@ -6,11 +6,12 @@ namespace FysioEnterprise.Domain.ValueObjects
 {
     public record Price
     {
-        public decimal Value { get; init; }
+        public double Value { get; init; }
         private Price() // Empty constructor for EF Core
         {
+
         }
-        public Price(decimal value)
+        public Price(double value)
         {
             if (value < 0) throw new ArgumentException("Amount cannot be negative.", nameof(value));
             Value = value;
